+++ src/dgisim/encoding/mappings.py
@@ -239,7 +239,7 @@ CARD_MAPPING: dict[type["Card"], int] = HashableDict({
         (card.SumeruCity, 259),
         (card.Vanarana, 260),
         # (card.ChinjuForest, 261),
-        # (card.GoldenHouse, 262),
+        (card.GoldenHouse, 262),
         (card.GandharvaVille, 263),
         # (card.StormterrorsLair, 264),
 
@@ -1006,7 +1006,7 @@ SUPP_MAPPING: dict[type["Support"], int] = HashableDict({
         (support.SumeruCitySupport, 9),
         (support.VanaranaSupport, 10),
         # (support.ChinjuForestSupport, 11),
-        # (support.GoldenHouseSupport, 12),
+        (support.GoldenHouseSupport, 12),
         (support.GandharvaVilleSupport, 13),
         # (support.StormterrorsLairSupport, 14),